src/pyproject.toml
[project]
name = "src"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "allure-pytest>=2.15.3",
    "pdoc3>=0.11.6",
    "prometheus-client>=0.23.1",
    "pytest>=9.0.2",
    "pyyaml>=6.0.3",
]
